Added junit test for create coding challenge
I added a junit test for create coding challenge, however I couldn't figure out how to make it in the test folder, and have it be able to access the coding challenge class. I fixed it by moving it into the src folder with the other files. The test runs and passes, but it might not be the most elegant setup to have it in src.